pyproject.toml
[project]
name = "condenser"
version = "0.1.0"
description = "Longitudinal stability benchmarking of quantum computing platforms"
requires-python = ">=3.11"
dependencies = [
    "amazon-braket-sdk>=1.88.0",
    "qiskit>=1.0.0",
    "qiskit-aqt-provider>=1.0.0",
    "pandas>=2.2.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-httpx>=0.36.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["archive/"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
filterwarnings = [
    # qiskit-aqt-provider uses deprecated Qiskit BackendV1 internals — not our code
    "ignore::DeprecationWarning:qiskit",
    "ignore::DeprecationWarning:qiskit_aer",
    "ignore::qiskit_aqt_provider.aqt_provider.NoTokenWarning",
    "ignore::pydantic.warnings.PydanticDeprecatedSince211",
]
